fix: Move plotter page to instance/id
