+++ backends-templates/backends.json
@@ -1,8 +1,11 @@
 {
-  "visualisationApiUrl":"{{env "VISUALISATION_API_URL"}}",
-  "discoveryApiUrl":"{{env "DISCOVERY_API_URL"}}",
-  "dssauthUrl":"{{env "DSS_AUTH"}}",
-  "node_env":"{{env "NODE_ENV"}}",
-  "instanceID":"{{env "INSTANCE_ID"}}",
-  "dataIds":{{include "/datasets/datasets.json"}}
+  "visualisationApiUrl": "{{env "VISUALISATION_API_URL"}}",
+  "discoveryApiUrl": "{{env "DISCOVERY_API_URL"}}",
+  "dssauthUrl": "{{env "DSS_AUTH"}}",
+  "node_env": "{{env "NODE_ENV"}}",
+  "instanceID": "{{env "INSTANCE_ID"}}",
+  "dataIds": {
+    {include "/datasets/datasets.json"
+    }
+  }
 }
